admin-panel-opencart/requirements.txt
# Основные зависимости
selenium==4.9.0
pytest==7.4.0

# Для отчетов Allure
allure-pytest==2.13.2
pytest-html==4.0.2

# Для генерации тестовых данных
faker==18.10.1

# Для логирования
pytest-logger==1.0.0

# Дополнительные утилиты
webdriver-manager==3.8.6  # Для автоматического управления драйверами браузеров